FST-6 added getting full info about film

+++ src/main/scala/by/bsu/utils/DirectorsFilmsService.scala
@@ -4,7 +4,7 @@ import by.bsu.model.dao.DirectorsFilmsDAO
 import by.bsu.model.repository.DirectorFilm
 
 class DirectorsFilmsService(directorsFilmsDAO: DirectorsFilmsDAO) {
-  def getByName(actorId: Int, filmId: Long) = {
+  def getByName(actorId: Int, filmId: Int) = {
     directorsFilmsDAO.findByName(actorId, filmId)
   }
 
@@ -16,7 +16,7 @@ class DirectorsFilmsService(directorsFilmsDAO: DirectorsFilmsDAO) {
     directorsFilmsDAO.insertListDirectorFilm(directorsFilmList)
   }
 
-  def deleteById(directorsId: Int, filmId: Long) = {
+  def deleteById(directorsId: Int, filmId: Int) = {
     directorsFilmsDAO.deleteById(directorsId, filmId)
   }
